pyproject.toml
[project]
name = "whichcat"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.42.91",
    "metaflow>=2.19.22",
    "mlflow>=3.11.1",
    "opencv-python>=4.13.0.92",
    "opencv-python-headless>=4.13.0.92",
    "pytest>=9.0.3",
    "python-dotenv>=1.2.2",
    "pyyaml>=6.0.3",
    "sagemaker>=3.8.0",
    "torch>=2.11.0",
    "torchaudio>=2.11.0",
    "torchvision>=0.26.0",
    "ultralytics>=8.4.40",
]
